LibWeb: Split ConicGradientStyleValue out of StyleValue.{h,cpp}
