+++ Userland/Libraries/LibWeb/Painting/GradientPainting.cpp
@@ -7,6 +7,7 @@
 #include <AK/Math.h>
 #include <LibGfx/Gradients.h>
 #include <LibWeb/CSS/StyleValue.h>
+#include <LibWeb/CSS/StyleValues/ConicGradientStyleValue.h>
 #include <LibWeb/Painting/GradientPainting.h>
 
 namespace Web::Painting {
